Add a chainguard_image_repos data source
This adds an image_repos (plural) data source next to the singular data source. Instead of expecting exactly one instance to be returned and failing if no instances can be returned, this follows the pattern of the AWS provider and creates a plural source that happily returns 0 or multiple instances.